refactor: move the WebPushEventType enum
