+++ Doppler.PushContact.WebPushSender.Test/Doppler.PushContact.WebPushSender.Test.csproj
@@ -17,6 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Doppler.PushContact.Models\Doppler.PushContact.Models.csproj" />
     <ProjectReference Include="..\Doppler.PushContact.WebPushSender\Doppler.PushContact.WebPushSender.csproj" />
   </ItemGroup>
