+++ Doppler.PushContact.WebPushSender/Senders/DefaultWebPushSender.cs
@@ -1,6 +1,6 @@
+using Doppler.PushContact.Models.Enums;
 using Doppler.PushContact.QueuingService.MessageQueueBroker;
 using Doppler.PushContact.WebPushSender.DTOs;
-using Doppler.PushContact.WebPushSender.Repositories;
 using Doppler.PushContact.WebPushSender.Repositories.Interfaces;
 using Doppler.PushContact.WebPushSender.Repositories.Models;
 using Microsoft.Extensions.Logging;
